requirements.txt
# OCR Runner Dependencies

# Core
requests>=2.28.0
Pillow>=9.0.0
loguru>=0.7.0

# PaddleOCR (install separately if needed)
# paddleocr>=2.7.0
# paddlepaddle>=2.5.0

# Comparison utilities (already in Utils/)
# No additional dependencies needed
